Code for Calculating Rate Statistics in Regions
Uploading a Jupyter notebook for calculating carbon sequestration rate statistics for regions such as ecozones and countries. Statistics include the annual average carbon sequestration rate for the region, lower and upper bounds of the confidence interval for the average rate in the region from our model variance, and the minimum and maximum observed rates in the region.